+++ src/nodes.py
@@ -8,6 +8,7 @@ from .generate_h3_prompts import GenerateH3PromptsNode
 from .load_chapter_catalogs import LoadChapterCatalogsNode
 from .load_consolidated_references import LoadConsolidatedReferencesNode
 from .lmstudio_config import LMStudioConfigurationNode
+from .spatial_continuity import SpatialContinuityNode
 
 
 NODE_CLASS_MAPPINGS = {
@@ -18,6 +19,7 @@ NODE_CLASS_MAPPINGS = {
     "LoadConsolidatedReferencesNode": LoadConsolidatedReferencesNode,
     "ConsolidateReferencesNode": ConsolidateReferencesNode,
     "GenerateH3PromptsNode": GenerateH3PromptsNode,
+    "SpatialContinuityNode": SpatialContinuityNode,
 }
 
 __all__ = [
@@ -28,5 +30,6 @@ __all__ = [
     "LoadConsolidatedReferencesNode",
     "ConsolidateReferencesNode",
     "GenerateH3PromptsNode",
+    "SpatialContinuityNode",
     "NODE_CLASS_MAPPINGS",
 ]